Avoid leaking internal module paths into repl API `.d.ts`
